requirements.txt
pandas
numpy
scikit-learn
scipy
pyod
tabulate
streamlit
great_expectations
openai
python-dotenv